+++ asapdiscovery-alchemy/MANIFEST.in
@@ -1,0 +1,5 @@
+include LICENSE
+include MANIFEST.in
+
+graft asapdiscovery
+global-exclude *.py[cod] __pycache__ *.so
