+++ asapdiscovery-alchemy/README.md
@@ -1,0 +1,3 @@
+# asapdiscovery-alchemy
+
+Tooling for performing alchemical free energy calculations via OpenFE and `alchemiscale`.
