+++ asapdiscovery-alchemy/asapdiscovery/alchemy/tests/test_fec_schema.py
@@ -1,18 +1,18 @@
 import openfe
 import pytest
 from alchemiscale import Scope, ScopedKey
-from asapdiscovery.simulation.alchemy.schema.atom_mapping import (
+from asapdiscovery.alchemy.schema.atom_mapping import (
     KartografAtomMapper,
     LomapAtomMapper,
     PersesAtomMapper,
 )
-from asapdiscovery.simulation.alchemy.schema.fec import (
+from asapdiscovery.alchemy.schema.fec import (
     AlchemiscaleResults,
     FreeEnergyCalculationFactory,
     SolventSettings,
     TransformationResult,
 )
-from asapdiscovery.simulation.alchemy.schema.network import NetworkPlanner
+from asapdiscovery.alchemy.schema.network import NetworkPlanner
 from openff.units import unit as OFFUnit
